added pagination and modified video controllers and db logic

+++ Services/DB_Services/VideoService.cs
@@ -2,6 +2,8 @@ using Microsoft.CodeAnalysis.CSharp.Syntax;
 using Microsoft.EntityFrameworkCore;
 using TaggerApi.DTOs;
 using TaggerApi.Models;
+using TaggerApi.Pagination;
+using TaggerApi.Extensions;
 using TaggerApi.Services.ErrorServices;
 
 namespace TaggerApi.Services.DB_Services;
@@ -10,19 +12,22 @@ public class VideoService : IVideoService
 {
     
     private readonly PostgresContext _context;
+    private readonly IPagedList _pageList;
 
-    public VideoService(PostgresContext context){
+    public VideoService(PostgresContext context, IPagedList pagination){
+        _pageList = pagination;
         _context = context;
     }
-
-    public async Task<VideoDTO> AddVideo(VideoDTO videoDTO)
+          
+    public async Task<VideoDTO> AddVideo(VideoDTO videoDTO,string userUid)
     {
+        
         var video = new Video{
             Id = videoDTO.Id,
             Name = videoDTO.Name,
             Link = videoDTO.Link,
             Description = videoDTO.Description,
-            IdUser = videoDTO.IdUser
+            IdUser = userUid
         };
             
         _context.Videos.Add(video);
@@ -31,13 +36,17 @@ public class VideoService : IVideoService
         return VideoToDTO(video);
     }
 
-    public async Task<bool> DelVideo(long id)
+    public async Task<bool> DelVideo(long id,string userUid)
     {
         var video = await _context.Videos.FindAsync(id);
         
         if (video == null)
         {
             return false;
+        }
+
+        if(userUid != video.IdUser){
+            throw new UnauthorizedAccessException("Video not your property.");
         }
 
         _context.Videos.Remove(video);
@@ -63,32 +72,34 @@ public class VideoService : IVideoService
         return await _context.Videos
                 .Select(x => VideoToDTO(x))
                 .ToListAsync();
-
     }
 
-    public async Task<VideoDTO> UpdateVideo(long id,VideoDTO videoDTO)
+    public async Task<IEnumerable<VideoDTO>> GetByUser(string userUid)
+    {
+        return await _context.Videos
+                .Where(b => b.IdUser == userUid)
+                .Select(x => VideoToDTO(x))
+                .ToListAsync();
+    }    
+
+    public async Task<VideoDTO> UpdateVideo(long id,VideoDTO videoDTO,string userUid)
     {
         var video = await _context.Videos.FindAsync(id);
-
+        
         if(video == null){
             throw new NotFoundException("Video not found.");
+        }
+
+        if(userUid != video.IdUser){
+            throw new UnauthorizedAccessException("Video not your property.");
         }
 
         video.Name = videoDTO.Name;
         video.Link = videoDTO.Link;
         video.Description = videoDTO.Description;
-        video.IdUser = videoDTO.IdUser;
 
-        try
-        {
-            await _context.SaveChangesAsync();
-        }
-        catch (DbUpdateConcurrencyException)
-        when (!VideoExists(id))
-        {
-           throw new NotFoundException("Video not found");
-        }
-
+        await _context.SaveChangesAsync();
+        
         return VideoToDTO(video);
     }
 
@@ -103,7 +114,28 @@ public class VideoService : IVideoService
         Id = video.Id,
         Name = video.Name,
         Link = video.Link,
-        Description = video.Description,
-        IdUser = video.IdUser
-    };    
+        Description = video.Description
+    };
+
+    public async Task<IEnumerable<VideoDTO>> GetByUserPag(PaginationParams request,string userUid)
+    {
+       var query =  _context.Videos
+                .Where(b => b.IdUser == userUid);
+                //.Select(x => VideoToDTO(x));
+
+       var resultPagination = await _pageList.CreatePagedGenericResults<Video>(query,
+           request.PageNumber,
+           request.PageSize,
+           request.OrderBy!,
+           request.OrderAsc
+       );
+
+       var listDto = new List<VideoDTO>();
+
+       foreach (var item in resultPagination.Result){
+            listDto.Add(VideoToDTO(item));
+       }
+
+       return listDto;
+    }
 }
